restructure file system
