+++ index.html
@@ -79,6 +79,6 @@
             </div>
         </div>       
     </div>
-<script src="./scripts/index.js" type="module"></script>
+<script src="./pages/index.js" type="module"></script>
 </body>
 </html>
