--- scripts/intialCards.js
@@ -1,26 +1,0 @@
-export const initialCards = [
-    {
-        name: 'Лондон',
-        link: './images/biel-morro-wydEbTWhzTc-unsplash.jpg',
-    },
-    {
-        name: 'Лувр',
-        link: './images/patrick-langwallner-TTz_H1FrdIc-unsplash.jpg',
-    },
-    {
-        name: 'Белладжо',
-        link: './images/luca-j-60LHJ-wdcgk-unsplash.jpg',
-    },
-    {
-        name: 'Москва-Сити',
-        link: './images/alexandr-bormotin-jj3PpeBIlLA-unsplash.jpg',
-    },
-    {
-        name: 'Морпорт Сочи',
-        link: './images/dima-fedorov-ZZkw00YCY1E-unsplash.jpg',
-    },
-    {
-        name: 'Санкт-Петербург',
-        link: './images/ilya-bronskiy-7haj8Ca19Ts-unsplash.jpg',
-    },
-];
